Implement importing of exported user profiles
This is a really unsafe and unreliable implementation but should be enough to share
configuration with others or backup them somwhere an load them later back into
the website.
Also the reset of the website should now contain one bug less by resetting the user profiles.
The visualizations are still not being reset because there is currently no use case for this
since you can just F5 the website after pressing the reset everything button.